% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHPDDP-package.R
\docType{package}
\name{MHPDDP-package}
\alias{MHPDDP}
\alias{MHPDDP-package}
\title{MHPDDP: MHP}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alex Z. Jiang \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
